Added Spawning, Time and Reset UI elements
Created buttons to change which object type is spawned, the timescale
and to reset the scene.
Increased the accuracy of the simulation and matched with slow
timescale.
Increased the number of concurrent object in the scene.
Changed from nearest to linear texture filtering.
Updated graphics to look better at higher resolutions.
Took new screenshot for documentation.